Support stdin binaries
Done in such a way as to minimise changes. The commandline used outside
of the config module now looks like "/path/to/binary < filepath"

The config module now regenerates the gdb script whenever a gdb command
is built, with the filepath included within the script.

The only change outside of the config module is to the Executor to
actually use stdin when running the binary outside of gdb.